Refactor revision ID retrieval functions and add new revids_bot module
