+++ include.php
@@ -11,6 +11,7 @@ include_once __DIR__ . '/bots/mdwiki_sql.php';
 
 include_once __DIR__ . '/bots/config.php';
 include_once __DIR__ . '/bots/helps.php';
+include_once __DIR__ . '/bots/revids_bot.php';
 include_once __DIR__ . '/bots/files_helps.php';
 include_once __DIR__ . '/bots/access_helps.php';
 include_once __DIR__ . '/bots/access_helps_new.php';
